+++ package.json
@@ -11,6 +11,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "aos": "^2.3.4",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
     "carousel": "0.0.3",
